Extracted UserSession and RegularStats stuff to separate files
